Deployments states fixes
